Lazily initialize MapData
This massively reduces the time it takes to open the galaxy type selection screen,
with it now being near-instant as opposed to the former good 10 seconds it took.

Do note that this does not touch the vanilla menu.

The APIs behind this are currently incubating - it's planned to merge them
into SLAPI once they have been bulletproofed enough